Reorganized fixture and updated imports
