+++ e2e/tests/admin/analytics/web-traffic.test.ts
@@ -1,4 +1,4 @@
-import {test, expect} from '../../../helpers/fixtures/ghost-instance';
+import {test, expect} from '../../../helpers/playwright';
 import {AnalyticsWebTrafficPage} from '../../../helpers/pages/admin';
 
 test.describe('Ghost Admin - Analytics Web Traffic', () => {
